+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -132,7 +132,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-28T09:40:18.790Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-29T07:52:33.432Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")