+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMSX27G3VH2S2JS32GPMKC7J"
+        "01KMW99M9WJBHJ8MBY9GPBY5JS"
     )
     r"""The unique identifier of the data source"""