+++ packages/orq-rc/RELEASES.md
@@ -6659,3 +6659,13 @@ Based on:
 - [python v4.7.0-rc.25] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.25 - packages/orq-rc
+
+## 2026-03-29 07:55:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.3 (2.869.25) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.26] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.26] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.26 - packages/orq-rc